// packages/payload/src/uploads/types.ts
export type UploadInstructionsRequest = {
  collectionSlug: UploadCollectionSlug
  docPrefix?: string
  filename: string
  filesize: number
  mimeType: string
}

export type UploadInstructions = {
  file: {
    filename: string
    mimeType: string
    size: number
    uploadReference: Record<string, unknown>
  }
} & (
  | {
      data?: unknown
      name: string
      type: 'dispatch'
    }
  | {
      request: {
        headers?: Record<string, string>
        method: 'POST' | 'PUT'
        url: string
      }
      type: 'http'
    }
)